add Symbol Group - Icons
add Icons symbol group, providing a selection of commonly used icons, which can be customised.

7 symbols provided in this group:

Cog
a complicated symbol that can create cog shapes based on given parameters to describe its teeth.

Download
a simple - but customisable - download icon, with an arrow and horizontal line. can be flipped to represent uploads, for example.

Home
a simple icon showing a home, complete with a roof and door. the propertions are customisable.

Pin
a simple symbol showing a commonly-used "location pin" shape. shape is customisable.

Clock
a visibly simple clock symbol. however, it is customisable from the time to the length and shape of its individual hands.

Bluetooth
a Bluetooth logo. this is a very rigid symbol - only its width ratio can be customised (other than its normal aspects e.g. colour, size). the default width multiplier should give a nice representation of the correct ratio when the symbol is a square (width = height).